+++ libs/common/api-interface/src/lib/category.interface.ts
@@ -1,0 +1,7 @@
+export interface ICategory {
+  title: string,
+  subCategory?: SubCategory
+}
+
+interface SubCategory extends ICategory{}
+
